Address feedback: move to base class, rename var and update test accordingly
